requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
mysqlclient==2.1.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
